+++ index.html
@@ -21,7 +21,10 @@
   <script src="/classe/pessoa/pessoa.js"></script>
   <script src="/classe/classeAbstrata/classeAbstrata.js"></script>
   <script src="/classe/desafios.js"></script> -->
+  <script src="/namespace/geometriaCirc.js"></script>
+  <script src="/namespace/geometriaRect.js"></script>
   <script src="/namespace/namespace.js"></script>
+  <script src="/modulos/modulos.js"></script>
 </body>
 
 </html>
