+++ namespace/geometriaCirc.js
@@ -1,0 +1,13 @@
+"use strict";
+var Geometria;
+(function (Geometria) {
+    let Area;
+    (function (Area) {
+        const PI = 3.14;
+        function circunferencia(raio) {
+            return PI * Math.pow(raio, 2);
+        }
+        Area.circunferencia = circunferencia;
+    })(Area = Geometria.Area || (Geometria.Area = {}));
+})(Geometria || (Geometria = {}));
+//# sourceMappingURL=geometriaCirc.js.map
