+++ namespace/geometriaRect.js
@@ -1,0 +1,12 @@
+"use strict";
+var Geometria;
+(function (Geometria) {
+    let Area;
+    (function (Area) {
+        function retangulo(base, altura) {
+            return base * altura;
+        }
+        Area.retangulo = retangulo;
+    })(Area = Geometria.Area || (Geometria.Area = {}));
+})(Geometria || (Geometria = {}));
+//# sourceMappingURL=geometriaRect.js.map
