+++ namespace/geometriaRect.ts
@@ -1,0 +1,10 @@
+namespace Geometria {
+
+        export namespace Area {
+   
+                export function retangulo( base: number, altura: number): number {
+                    return base * altura
+                }
+        }
+    }
+
